hard break
