+++ pkg/loki/config_test.go
@@ -1,85 +1,132 @@
 package loki
 
 import (
+	"fmt"
+	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v2"
 )
 
-func TestConfigVersioning(t *testing.T) {
-	expectStr := `
-  version: v1
-  configs:
-  - name: default
-    positions:
-      filename: /tmp/positions.yml
-    clients:
-      - url: http://loki:3100/loki/api/v1/push
-    scrape_configs:
-    - job_name: varlog
-      static_configs:
-        - targets: [localhost]
-          labels:
-            job: varlog
-            __path__: /var/log/*log
-  `
-	var expect LatestConfig
-	err := yaml.UnmarshalStrict([]byte(expectStr), &expect)
-	require.NoError(t, err)
-
-	input := map[string]string{
-		"unversioned": `
-      positions:
-        filename: /tmp/positions.yml
-      clients:
-        - url: http://loki:3100/loki/api/v1/push
-      scrape_configs:
-      - job_name: varlog
-        static_configs:
-          - targets: [localhost]
-            labels:
-              job: varlog
-              __path__: /var/log/*log
-    `,
-		"v0": `
-     version: v0
-     positions:
-       filename: /tmp/positions.yml
-     clients:
-       - url: http://loki:3100/loki/api/v1/push
-     scrape_configs:
-      - job_name: varlog
-        static_configs:
-          - targets: [localhost]
-            labels:
-              job: varlog
-              __path__: /var/log/*log
-    `,
-		"v1": `
-      version: v1
-      configs:
-      - name: default
-        positions:
-          filename: /tmp/positions.yml
-        clients:
-          - url: http://loki:3100/loki/api/v1/push
-        scrape_configs:
-        - job_name: varlog
-          static_configs:
-            - targets: [localhost]
-              labels:
-                job: varlog
-                __path__: /var/log/*log
-    `,
+func TestConfig_ApplyDefaults_Validations(t *testing.T) {
+	tt := []struct {
+		name string
+		cfg  string
+		err  error
+	}{
+		{
+			name: "two configs with different names",
+			err:  nil,
+			cfg: untab(`
+				positions_directory: /tmp
+				configs: 
+				- name: config-a
+				- name: config-b
+		  `),
+		},
+		{
+			name: "two configs with same name",
+			err:  fmt.Errorf("found two Loki configs with name config-a"),
+			cfg: untab(`
+				positions_directory: /tmp
+				configs: 
+				- name: config-a
+				- name: config-b
+				- name: config-a
+		  `),
+		},
+		{
+			name: "two configs, different positions path",
+			err:  nil,
+			cfg: untab(`
+				configs: 
+				- name: config-a
+				  positions:
+					  filename: /tmp/file-a.yml
+				- name: config-b
+				  positions:
+					  filename: /tmp/file-b.yml
+		  `),
+		},
+		{
+			name: "re-used positions path",
+			err:  fmt.Errorf("Loki configs config-a and config-c must have different positions file paths"),
+			cfg: untab(`
+				configs: 
+				- name: config-a
+				  positions:
+					  filename: /tmp/file-a.yml
+				- name: config-b
+				  positions:
+					  filename: /tmp/file-b.yml
+				- name: config-c
+				  positions:
+					  filename: /tmp/file-a.yml
+		  `),
+		},
+		{
+			name: "empty name",
+			err:  fmt.Errorf("Loki config index 1 must have a name"),
+			cfg: untab(`
+				positions_directory: /tmp
+				configs: 
+				- name: config-a
+				- name:
+				- name: config-a
+		  `),
+		},
+		{
+			name: "generated positions file path without positions_directory",
+			err:  fmt.Errorf("cannot generate Loki positions file path for config-b because positions_directory is not configured"),
+			cfg: untab(`
+				configs: 
+				- name: config-a
+				  positions:
+					  filename: /tmp/config-a.yaml
+				- name: config-b
+		  `),
+		},
 	}
 
-	for name, ii := range input {
-		t.Run(name, func(t *testing.T) {
+	for _, tc := range tt {
+		t.Run(tc.name, func(t *testing.T) {
 			var cfg Config
-			err := yaml.UnmarshalStrict([]byte(ii), &cfg)
-			require.NoError(t, err)
-			require.Equal(t, expect, cfg.Config)
+			err := yaml.UnmarshalStrict([]byte(tc.cfg), &cfg)
+			if tc.err == nil {
+				require.NoError(t, err)
+			} else {
+				require.EqualError(t, err, tc.err.Error())
+			}
 		})
 	}
+}
+
+func TestConfig_ApplyDefaults_Defaults(t *testing.T) {
+	cfgText := untab(`
+		positions_directory: /tmp
+		configs:
+		- name: config-a
+			positions:
+				filename: /config-a.yml
+		- name: config-b
+	`)
+	var cfg Config
+	err := yaml.UnmarshalStrict([]byte(cfgText), &cfg)
+	require.NoError(t, err)
+
+	var (
+		pathA = cfg.Configs[0].PositionsConfig.PositionsFile
+		pathB = cfg.Configs[1].PositionsConfig.PositionsFile
+	)
+
+	require.Equal(t, "/config-a.yml", pathA)
+	require.Equal(t, filepath.Join("/tmp", "config-b.yml"), pathB)
+}
+
+// untab is a utility function to make it easier to write YAML tests, where some editors
+// will insert tabs into strings by default.
+func untab(s string) string {
+	return strings.ReplaceAll(s, "\t", "  ")
 }
